Add flag --min-aln-len

+++ src/commons/FileUtil.cpp
@@ -11,6 +11,7 @@
 #include <sys/types.h>
 #include <dirent.h>
 #include <sys/mman.h>
+#include <fts.h>
 
 bool FileUtil::fileExists(const char* fileName) {
     struct stat st;
@@ -83,6 +84,71 @@ size_t FileUtil::countLines(const char* name) {
     return cnt;
 }
 
+int FileUtil::recursiveDelete(const char *dir){
+
+    int ret = 0;
+    FTS *ftsp = NULL;
+    FTSENT *curr;
+
+    // Cast needed (in C) because fts_open() takes a "char * const *", instead
+    // of a "const char * const *", which is only allowed in C++. fts_open()
+    // does not modify the argument.
+    char *files[] = { (char *) dir, NULL };
+
+    // FTS_NOCHDIR  - Avoid changing cwd, which could cause unexpected behavior
+    //                in multithreaded programs
+    // FTS_PHYSICAL - Don't follow symlinks. Prevents deletion of files outside
+    //                of the specified directory
+    // FTS_XDEV     - Don't cross filesystem boundaries
+    ftsp = fts_open(files, FTS_NOCHDIR | FTS_PHYSICAL | FTS_XDEV, NULL);
+    if (!ftsp) {
+        fprintf(stderr, "%s: fts_open failed: %s\n", dir, strerror(errno));
+        ret = -1;
+        goto finish;
+    }
+
+    while ((curr = fts_read(ftsp))) {
+        switch (curr->fts_info) {
+            case FTS_NS:
+            case FTS_DNR:
+            case FTS_ERR:
+                Debug(Debug::ERROR) << curr->fts_accpath << ": fts_read error: " << strerror(curr->fts_errno) << "\n";
+                break;
+
+            case FTS_DC:
+            case FTS_DOT:
+            case FTS_NSOK:
+                // Not reached unless FTS_LOGICAL, FTS_SEEDOT, or FTS_NOSTAT were
+                // passed to fts_open()
+                break;
+
+            case FTS_D:
+                // Do nothing. Need depth-first search, so directories are deleted
+                // in FTS_DP
+                break;
+
+            case FTS_DP:
+            case FTS_F:
+            case FTS_SL:
+            case FTS_SLNONE:
+            case FTS_DEFAULT:
+                if (remove(curr->fts_accpath) < 0) {
+                    Debug(Debug::ERROR) << curr->fts_path << ": Failed to remove: " <<  strerror(errno) << "\n";
+                    ret = -1;
+                }
+                break;
+        }
+    }
+
+    finish:
+    if (ftsp) {
+        fts_close(ftsp);
+    }
+
+    return ret;
+
+}
+
 void FileUtil::deleteFile(const std::string &file) {
     if (remove(file.c_str()) != 0) {
         Debug(Debug::WARNING) << "Error deleting file " << file << "\n";
@@ -116,10 +182,10 @@ void FileUtil::writeFile(const std::string &pathToFile, const unsigned char *dat
 }
 
 std::string FileUtil::dirName(const std::string &file) {
-        size_t pos = file.find_last_of("\\/");
-        return (std::string::npos == pos)
-               ? "."
-               : file.substr(0, pos);
+    size_t pos = file.find_last_of("\\/");
+    return (std::string::npos == pos)
+           ? "."
+           : file.substr(0, pos);
 }
 
 std::string FileUtil::baseName(const std::string &file) {
@@ -130,14 +196,14 @@ std::string FileUtil::baseName(const std::string &file) {
 }
 
 size_t FileUtil::getFreeSpace(const char *path) {
-        struct statvfs stat;
-        if (statvfs(path, &stat) != 0) {
-            // error happens, just quits here
-            return -1;
-        }
+    struct statvfs stat;
+    if (statvfs(path, &stat) != 0) {
+        // error happens, just quits here
+        return -1;
+    }
 
-        // the available size is f_bsize * f_bavail
-        return stat.f_bfree * stat.f_frsize;
+    // the available size is f_bsize * f_bavail
+    return stat.f_bfree * stat.f_frsize;
 }
 
 std::string FileUtil::getHashFromSymLink(const std::string path){
@@ -260,7 +326,13 @@ void FileUtil::copyFile(const char *src, const char *dst) {
 
 FILE * FileUtil::openAndDelete(const char *fileName, const char *mode) {
     if(FileUtil::fileExists(fileName) == true){
-        FileUtil::deleteFile(fileName);
+        if(FileUtil::directoryExists(fileName)){
+            Debug(Debug::ERROR) << "Can not open " << fileName << " for writing. It is a directory.\n";
+            EXIT(EXIT_FAILURE);
+//            FileUtil::recursiveDelete(fileName);
+        }else {
+            FileUtil::deleteFile(fileName);
+        }
     }
     FILE * file = fopen(fileName, mode);
     if (file == NULL) {
